Update explorer.md file
This PR updates the explorer.md file with the necessary changes to reflect the latest information and instructions.

The modifications include adding a new step after step 4 in the deployment document. Specifically, it instructs users to edit docker-compose-no-build-geth.yml and replace blockscout/blockscout:latest with blockscout/blockscout-optimism:latest. This change ensures that Blockscout is configured correctly to work with the Optimism network.

Please review and merge this PR. Thank you!