Update data-avail.md: Non-EVM DA (PoDA)
Please read!  Your attention is appreciated.

The first Non-EVM DA within the OP Stack Community exists, and it's a protocol called Proof of Data Availability.  It is used by the Rollux fork of Optimism.  PoDA is L1 and has been mainnet for about eight months and Rollux has been mainnet since June 28th.

You can find out about PoDA's design here:
https://docs.rollux.com/docs/sys/PoDA/#optimistic-rollup-with-poda-proof-of-data-availability

PoDA is available for anyone to use. Being PoW at its base with a hybrid finality makes it an interesting Multi-DA option in particular if a Rollup doesn't want to depend 100% on a single Proof-of-Stake based DA solution (which seem to make up the overwhelming majority).  Good approach for architectural redundancy.  PoDA does not use sharding.  State proof is settled on a UTXO-based chain that is merge mined with Bitcoin, and raw data is archived offchain in its entirety by each participating full node (open network), bringing a lot of censorship resistance.

To see the DA difference in the code, a delta between the OP and Rollux approach to DA is visible here:
SYS-Labs/rollux@25a4c94

ALSO, since OP is pursuing modularization of DA in Superchain, it's worth mentioning that the Rollux Team and Syscoin Foundation have some interest in contributing there.

All the best!